use slices for advertising sets
This allows dynamically passing less than the max sets if necessary,
which can be the case if the underlying controller has limits on
how many connections it can establish.

The price to pay is that the api for doing extended advertisements
becomes a bit harder to use, but some convenience fn should improve
that.